+++ bikespace_frontend/src/components/shared/base-button/base-button.module.scss
@@ -1,0 +1,49 @@
+@import '@styles/variables';
+
+.baseButton {
+  position: relative;
+  display: flex;
+  align-items: center;
+  background-color: white;
+  color: $secondary-color;
+  border: 2px solid $secondary-color;
+  height: 44px;
+  width: 100%;
+  border-radius: 23px;
+  padding-left: 10px;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: all 0.2s ease-in-out;
+
+  .check {
+    position: absolute;
+    right: 14px;
+    top: 50%;
+    opacity: 0;
+    transform: translateY(-50%);
+  }
+
+  &:focus,
+  &:focus-within {
+    box-shadow: 0 0 0px 4px $main-color;
+  }
+
+  &:focus:active {
+    box-shadow: none;
+  }
+}
+
+input[type='radio']:checked + label .baseButton,
+input[type='checkbox']:checked + label .baseButton {
+  background-color: $secondary-color;
+  color: white;
+
+  .check {
+    opacity: 1;
+  }
+}
+
+input[type='radio']:focus + label .baseButton,
+input[type='checkbox']:focus + label .baseButton {
+  box-shadow: 0 0 0px 4px $main-color;
+}
