+++ bikespace_frontend/src/components/submission/comments/comments.module.scss
@@ -1,0 +1,15 @@
+.submissionComments {
+  textarea {
+    display: flex;
+    margin-left: auto;
+    margin-right: auto;
+    border-radius: 10px;
+    padding: 10px;
+    resize: none;
+    width: 100%;
+    font-size: 1rem;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+      Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
+      'Segoe UI Symbol';
+  }
+}
