+++ bikespace_frontend/src/components/submission/issue/issue.module.scss
@@ -1,0 +1,19 @@
+.submissionIssue {
+  fieldset {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    margin: 0;
+    padding: 5px;
+    border: none;
+  }
+
+  ul {
+    padding: 5px;
+
+    li {
+      list-style: none;
+      margin-bottom: 20px;
+    }
+  }
+}
