+++ bikespace_frontend/src/components/submission/location/location.module.scss
@@ -1,0 +1,11 @@
+.location {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+
+  .outerMapContainer {
+    position: relative;
+    flex-grow: 1;
+    margin-bottom: 25px;
+  }
+}
