+++ bikespace_frontend/src/components/submission/submission-progress/SubmissionProgressBar.tsx
@@ -8,14 +8,16 @@ type SubmissionProgressBarProps = {
 
 export function SubmissionProgressBar({step}: SubmissionProgressBarProps) {
   return (
-    <div className={styles.progressBar}>
-      <div className={styles.middleLine} />
-      {[0, 1, 2, 3, 4].map(s => (
-        <div
-          key={s}
-          className={`${styles.step} ${s === step ? styles.active : ''}`}
-        />
-      ))}
-    </div>
+    <header>
+      <div className={styles.progressBar}>
+        <div className={styles.middleLine} />
+        {[0, 1, 2, 3, 4].map(s => (
+          <div
+            key={s}
+            className={`${styles.step} ${s === step ? styles.active : ''}`}
+          />
+        ))}
+      </div>
+    </header>
   );
 }
