+++ bikespace_frontend/src/components/submission/time/time.module.scss
@@ -1,0 +1,24 @@
+@import '@styles/variables';
+
+.submissionTime {
+  fieldset > .checkbox-group {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 14px 0;
+  }
+
+  input[type='datetime-local'] {
+    width: 100%;
+    border: none;
+    border-radius: 21px;
+    margin: 0px;
+    margin-bottom: 10px;
+    padding: 10px;
+    border: 2px solid $gray-color;
+    text-align: left;
+    font-size: 1rem;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+      Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
+      'Segoe UI Symbol';
+  }
+}
