+++ bikespace_frontend/src/styles/global.scss
@@ -114,51 +114,3 @@ input[type='radio'] {
     display: none;
   }
 }
-
-input[type='radio']:checked + label .base-button,
-input[type='checkbox']:checked + label .base-button {
-  background-color: $secondary-color;
-  color: white;
-
-  .check {
-    opacity: 1;
-  }
-}
-
-.base-button {
-  position: relative;
-  display: flex;
-  align-items: center;
-  background-color: white;
-  color: $secondary-color;
-  border: 2px solid $secondary-color;
-  height: 44px;
-  width: 100%;
-  border-radius: 23px;
-  padding-left: 10px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-
-  .check {
-    position: absolute;
-    right: 14px;
-    top: 50%;
-    opacity: 0;
-    transform: translateY(-50%);
-  }
-
-  &:focus,
-  &:focus-within {
-    box-shadow: 0 0 0px 4px $main-color;
-  }
-
-  &:focus:active {
-    box-shadow: none;
-  }
-}
-
-input[type='radio']:focus + label .base-button,
-input[type='checkbox']:focus + label .base-button {
-  box-shadow: 0 0 0px 4px $main-color;
-}
