+++ bikespace_frontend/src/styles/submission.scss
@@ -3,7 +3,7 @@
 
 #submission {
   position: relative;
-  
+
   main {
     position: relative;
     padding: 20px;
@@ -61,113 +61,5 @@
       display: flex;
       justify-content: space-between;
     }
-  }
-}
-
-#submission-progress-bar {
-  position: relative;
-  display: flex;
-  justify-content: space-between;
-  width: 90%;
-  margin: 0 auto;
-  padding: 16px 0;
-
-  #middle-line {
-    position: absolute;
-    top: 50%;
-    left: 0;
-    transform: translateY(-50%);
-    width: 100%;
-    height: 2px;
-    background-color: $gray-color;
-  }
-
-  .step {
-    width: 12px;
-    height: 12px;
-    border-radius: 50%;
-    background-color: $gray-color;
-    transition: background-color 0.3s ease-in-out;
-    z-index: 10;
-
-    &.active {
-      background-color: $main-color;
-    }
-  }
-}
-
-#submission-issue {
-  fieldset {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    margin: 0;
-    padding: 5px;
-    border: none;
-  }
-
-  ul {
-    padding: 5px;
-
-    li {
-      list-style: none;
-      margin-bottom: 20px;
-    }
-  }
-}
-
-#submission-location {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-
-  #outer-map-container {
-    position: relative;
-    flex-grow: 1;
-    margin-bottom: 25px;
-  }
-}
-
-#submission-time {
-  fieldset > .checkbox-group {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 14px 0;
-
-    .base-button {
-      margin: 0 7px;
-      width: revert;
-    }
-  }
-
-  input[type='datetime-local'] {
-    width: 100%;
-    border: none;
-    border-radius: 21px;
-    margin: 0px;
-    margin-bottom: 10px;
-    padding: 10px;
-    border: 2px solid $gray-color;
-    text-align: left;
-    font-size: 1rem;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-      Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-      'Segoe UI Symbol';
-  }
-}
-
-#submission-comments {
-  textarea {
-    display: flex;
-    margin-left: auto;
-    margin-right: auto;
-    border-radius: 10px;
-    padding: 10px;
-    resize: none;
-    width: 100%;
-    font-size: 1rem;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-      Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-      'Segoe UI Symbol';
   }
 }
